pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Michail Fedyaev <drovosekov1910@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
ipython = "^8.28.0"
telethon = "^1.37.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
